requirements.txt
absl-py==0.7.1
APScheduler==3.6.0
asn1crypto==0.24.0
astor==0.8.0
attrs==19.1.0
Automat==0.7.0
blis==0.2.4
boto==2.49.0
boto3==1.9.175
botocore==1.12.175
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
cloudpickle==0.6.1
colorama==0.4.1
colorclass==2.2.0
coloredlogs==10.0
colorhash==1.0.2
ConfigArgParse==0.13.0
constantly==15.1.0
cryptography==2.7
cycler==0.10.0
cymem==2.0.2
decorator==4.4.0
Django==2.2.2
djangorestframework==3.9.4
docopt==0.6.2
docutils==0.14
#en-core-web-sm==2.1.0
fakeredis==0.10.3
fbmessenger==5.6.0
Flask==1.0.3
Flask-Cors==3.0.8
Flask-JWT-Simple==0.0.3
future==0.17.1
gast==0.2.2
gensim==3.7.3
gevent==1.4.0
google-pasta==0.1.7
greenlet==0.4.15
grpcio==1.21.1
h5py==2.9.0
humanfriendly==4.18
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
inflect==2.1.0
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
jsonpickle==1.2
jsonschema==2.6.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
klein==17.10.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==2.2.4
mattermostwrapper==2.1
mock==3.0.5
murmurhash==1.0.2
networkx==2.3
nltk==3.4.3
numpy==1.16.4
packaging==18.0
pandas==0.24.2
pika==0.12.0
plac==0.9.6
preshed==2.0.1
prompt-toolkit==2.0.9
protobuf==3.8.0
pycparser==2.19
pydot==1.4.1
PyHamcrest==1.9.0
PyJWT==1.7.1
pykwalify==1.7.0
pymongo==3.8.0
pyparsing==2.4.0
pyreadline==2.1
PySocks==1.7.0
pyspellchecker==0.4.0
python-crfsuite==0.9.6
python-dateutil==2.8.0
python-engineio==3.8.1
python-socketio==3.1.2
python-telegram-bot==11.1.0
pytz==2018.9
PyYAML==5.1.1
questionary==1.1.1
rasa-core==0.14.5
rasa-core-sdk==0.14.0
rasa-nlu==0.15.1
redis==2.10.6
requests==2.22.0
requests-toolbelt==0.9.1
rocketchat-API==0.6.32
ruamel.yaml==0.15.97
s3transfer==0.2.1
scikit-learn==0.20.3
scipy==1.3.0
seaborn==0.9.0
simplejson==3.16.0
six==1.12.0
sklearn==0.0
sklearn-crfsuite==0.3.6
slackclient==1.3.2
smart-open==1.8.4
spacy==2.1.4
sqlparse==0.3.0
srsly==0.0.7
tabulate==0.8.3
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
#tensorflow-gpu==1.14.0
termcolor==1.1.0
terminaltables==3.1.0
thinc==7.0.4
tqdm==4.32.2
twilio==6.28.0
Twisted==15.5.0
typing==3.7.4
tzlocal==1.5.1
urllib3==1.25.3
wasabi==0.2.2
wcwidth==0.1.7
webexteamssdk==1.1.1
websocket-client==0.54.0
Werkzeug==0.15.4
wrapt==1.11.2
zope.interface==4.6.0
